post and category commands all in one
