+++ Dayana/Shared/Persistence/Models/Blog/Commands/PostCategoryCommands/CreatePostCategoryCommand.cs
@@ -5,17 +5,7 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.PostCategoryCommands;
 
-public class CreatePostCategoryCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public CreatePostCategoryCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public string CategoryTitle { get; set; }
-    public string CategoryIcon { get; set; }
 
-    public RequestInfo RequestInfo { get; set; }
-}
 
 public class CreatePostCategoryCommandValidator : AbstractValidator<CreatePostCategoryCommand>
 {
