+++ Dayana/Shared/Persistence/Models/Blog/Commands/PostCategoryCommands/DeletePostCategoryCommand.cs
@@ -5,15 +5,7 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.PostCategoryCommands;
 
-public class DeletePostCategoryCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public DeletePostCategoryCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public int Id { get; set; }
-    public RequestInfo RequestInfo { get; set; }
-}
+
 
 public class DeletePostCategoryCommandValidator : AbstractValidator<DeletePostCategoryCommand>
 {
