+++ Dayana/Shared/Persistence/Models/Blog/Commands/PostCategoryCommands/UpdatePostCategoryCommand.cs
@@ -5,18 +5,7 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.PostCategoryCommands;
 
-public class UpdatePostCategoryCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public UpdatePostCategoryCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public int Id { get; set; }
-    public string CategoryTitle { get; set; }
-    public string CategoryIcon { get; set; }
 
-    public RequestInfo RequestInfo { get; set; }
-}
 
 public class UpdatePostCategoryCommandValidator : AbstractValidator<UpdatePostCategoryCommand>
 {
