+++ Dayana/Shared/Persistence/Models/Blog/Commands/PostCategoryCommandsFile.cs
@@ -1,0 +1,44 @@
+﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
+using Dayana.Shared.Infrastructure.Operations;
+using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Dayana.Shared.Persistence.Models.Blog.Commands;
+public class CreatePostCategoryCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public CreatePostCategoryCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public string CategoryTitle { get; set; }
+    public string CategoryIcon { get; set; }
+
+    public RequestInfo RequestInfo { get; set; }
+}
+
+public class DeletePostCategoryCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public DeletePostCategoryCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public int Id { get; set; }
+    public RequestInfo RequestInfo { get; set; }
+}
+
+public class UpdatePostCategoryCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public UpdatePostCategoryCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public int Id { get; set; }
+    public string CategoryTitle { get; set; }
+    public string CategoryIcon { get; set; }
+
+    public RequestInfo RequestInfo { get; set; }
+}
