+++ Dayana/Shared/Persistence/Models/Blog/Commands/PostCommands/DeletePostCommand.cs
@@ -5,16 +5,7 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.PostCommands;
 
-public class DeletePostCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public RequestInfo RequestInfo { get; set; }
 
-    public DeletePostCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-    public int Id { get; set; }
-}
 
 public class DeletePostCommandValidator : AbstractValidator<DeletePostCommand>
 {
