+++ Dayana/Shared/Persistence/Models/Blog/Commands/PostCommands/UpdatePostCommand.cs
@@ -5,20 +5,7 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Commands.Blog.PostCommands;
 
-public class UpdatePostCommand : IRequestInfo, IRequest<OperationResult>
-{
-    public RequestInfo RequestInfo { get; set; }
 
-    public UpdatePostCommand(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-
-    public int Id { get; set; }
-    public string Title { get; set; }
-    public string Summery { get; set; }
-    public string TextContent { get; set; }
-}
 
 public class UpdatePostCommandValidator : AbstractValidator<UpdatePostCommand>
 {
