+++ Dayana/Shared/Persistence/Models/Blog/Commands/PostCommandsFile.cs
@@ -1,0 +1,49 @@
+﻿using Dayana.Shared.Basic.MethodsAndObjects.Models;
+using Dayana.Shared.Infrastructure.Operations;
+using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Dayana.Shared.Persistence.Models.Blog.Commands;
+
+public class CreatePostCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public CreatePostCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+
+    public string Title { get; set; }
+    public string Summery { get; set; }
+    public string TextContent { get; set; }
+    public RequestInfo RequestInfo { get; private set; }
+}
+
+public class DeletePostCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public RequestInfo RequestInfo { get; set; }
+
+    public DeletePostCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+    public int Id { get; set; }
+}
+
+public class UpdatePostCommand : IRequestInfo, IRequest<OperationResult>
+{
+    public RequestInfo RequestInfo { get; set; }
+
+    public UpdatePostCommand(RequestInfo requestInfo)
+    {
+        RequestInfo = requestInfo;
+    }
+
+    public int Id { get; set; }
+    public string Title { get; set; }
+    public string Summery { get; set; }
+    public string TextContent { get; set; }
+}
